backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
motor>=3.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart
pyjwt
passlib[bcrypt]
tensorflow
numpy
pandas
requests
python-dotenv
joblib
scikit-learn
